requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
rdkit>=2023.3.1
biopython>=1.81
meeko>=0.4.0
scipy>=1.10.0
seaborn>=0.12.2
triton; platform_system=='Linux'
py3Dmol>=2.0.3
pandas>=2.0.0
matplotlib>=3.7.1
